feat: ✨ useStockChangeLabel 훅으로 분리
